data_pipeline_bbox_only/requirements.txt
numpy>=1.20
torch>=2.0
torch-geometric>=2.0
scikit-learn>=1.0
tqdm>=4.60
PyYAML>=6.0
